pr review comments implementation
