+++ app/src/androidTest/java/com/swent/assos/CommitteeDetailsTest.kt
@@ -4,16 +4,13 @@ import androidx.activity.compose.setContent
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.isDisplayed
 import androidx.compose.ui.test.onNodeWithText
-import androidx.compose.ui.test.performClick
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.swent.assos.model.data.CommitteeMember
 import com.swent.assos.model.data.User
 import com.swent.assos.model.serialize
-import com.swent.assos.screens.AssoDetailsScreen
 import com.swent.assos.screens.CreateNewsScreen
-import com.swent.assos.ui.screens.assoDetails.AssoDetails
 import com.swent.assos.ui.screens.assoDetails.CommitteeDetails
 import dagger.hilt.android.testing.HiltAndroidTest
 import io.github.kakaocup.compose.node.element.ComposeScreen
@@ -46,24 +43,6 @@ class CommitteeDetailsTest : SuperTest() {
         .collection("committee")
         .document(member.id)
         .set(member)
-  }
-
-  @Test
-  fun testNavigationToCommitteeDetails() {
-    composeTestRule.activity.setContent {
-      AssoDetails(navigationActions = mockNavActions, assoId = assoId)
-    }
-
-    run {
-      ComposeScreen.onComposeScreen<AssoDetailsScreen>(composeTestRule) {
-        step("Check if the committee button is displayed") {
-          composeTestRule.onNodeWithText("The Committee").assertIsDisplayed()
-        }
-        step("Click on the committee button") {
-          composeTestRule.onNodeWithText("The Committee").performClick()
-        }
-      }
-    }
   }
 
   @Test
