+++ app/src/main/java/com/swent/assos/ui/screens/assoDetails/AssoDetails.kt
@@ -4,7 +4,6 @@ import android.net.Uri
 import android.widget.Toast
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -16,16 +15,16 @@ import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Accessibility
 import androidx.compose.material3.AssistChip
 import androidx.compose.material3.AssistChipDefaults
+import androidx.compose.material3.Button
 import androidx.compose.material3.FabPosition
-import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.FloatingActionButtonDefaults
+import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.MediumTopAppBar
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -36,14 +35,12 @@ import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTagsAsResourceId
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
@@ -52,10 +49,9 @@ import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import coil.compose.rememberAsyncImagePainter
 import com.swent.assos.R
-import com.swent.assos.model.data.Association
+import com.swent.assos.model.navigation.Destinations
 import com.swent.assos.model.navigation.NavigationActions
 import com.swent.assos.model.view.AssoViewModel
-import com.swent.assos.model.view.ProfileViewModel
 import com.swent.assos.ui.components.EventItem
 import com.swent.assos.ui.components.NewsItem
 import com.swent.assos.ui.components.PostionItem
@@ -75,8 +71,6 @@ fun AssoDetails(assoId: String, navigationActions: NavigationActions) {
 
   val listStateNews = rememberLazyListState()
   val listStateEvents = rememberLazyListState()
-
-  val listStatePos = rememberLazyListState()
 
   val context = LocalContext.current
 
@@ -99,7 +93,7 @@ fun AssoDetails(assoId: String, navigationActions: NavigationActions) {
   LaunchedEffect(listStateEvents) {
     snapshotFlow { listStateEvents.layoutInfo.visibleItemsInfo }
         .collect { visibleItems ->
-          if (visibleItems.isNotEmpty() && visibleItems.last().index == events.size - 1) {
+          if (visibleItems.lastOrNull()?.index == events.size - 1) {
             viewModel.getMoreEvents(assoId)
           }
         }
@@ -254,6 +248,20 @@ fun AssoDetails(assoId: String, navigationActions: NavigationActions) {
       }
 
       item {
+        Button(
+            modifier =
+                Modifier.testTag("CommitteeButton").padding(30.dp).width(250.dp).height(60.dp),
+            shape = RoundedCornerShape(16.dp),
+            onClick = {
+              navigationActions.navigateTo(
+                  Destinations.COMMITTEE_DETAILS.route + "/${assoId}" + "/${association.acronym}")
+            }) {
+              Icon(imageVector = Icons.Default.Accessibility, contentDescription = null)
+              Text("The Committee", modifier = Modifier.padding(start = 12.dp), fontSize = 18.sp)
+            }
+      }
+
+      item {
         Text(
             text = "Latest Positions",
             style = MaterialTheme.typography.headlineMedium,
@@ -278,112 +286,7 @@ fun AssoDetails(assoId: String, navigationActions: NavigationActions) {
                   modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp))
         }
       }
-
-      item {
-        Text(
-            text = "The Committee",
-            style = MaterialTheme.typography.headlineMedium,
-            fontFamily = FontFamily(Font(R.font.sf_pro_display_regular)),
-            fontWeight = FontWeight.Bold,
-            modifier =
-                Modifier.padding(horizontal = 16.dp, vertical = 10.dp).clickable {
-                  navigationActions.navigateTo(
-                      Destinations.COMMITTEE_DETAILS.route +
-                          "/${assoId}" +
-                          "/${association.acronym}")
-                })
-        Spacer(modifier = Modifier.height(50.dp))
-      }
-      item { Spacer(modifier = Modifier.height(20.dp)) }
+      item { Spacer(modifier = Modifier.height(60.dp)) }
     }
-  }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun TopAssoBar(asso: Association, navigationActions: NavigationActions, viewModel: AssoViewModel) {
-  val associationFollowed = viewModel.associationFollowed.collectAsState()
-  val profileViewModel: ProfileViewModel = hiltViewModel()
-
-  MediumTopAppBar(
-      colors = TopAppBarDefaults.topAppBarColors(containerColor = Color.Transparent),
-      modifier = Modifier.testTag("Header"),
-      title = {
-        Text(
-            asso.acronym,
-            modifier = Modifier.testTag("Title"),
-            style =
-                TextStyle(
-                    fontSize = 30.sp,
-                    lineHeight = 32.sp,
-                    fontFamily = FontFamily(Font(R.font.sf_pro_display_regular)),
-                    fontWeight = FontWeight.SemiBold,
-                    color = MaterialTheme.colorScheme.onBackground))
-      },
-      navigationIcon = {
-        Image(
-            colorFilter = ColorFilter.tint(MaterialTheme.colorScheme.onBackground),
-            imageVector = Icons.Default.ArrowBack,
-            contentDescription = null,
-            modifier = Modifier.testTag("GoBackButton").clickable { navigationActions.goBack() })
-      },
-      actions = {
-        if (asso.id != "") {
-          AssistChip(
-              colors =
-                  if (associationFollowed.value)
-                      AssistChipDefaults.assistChipColors(
-                          containerColor = MaterialTheme.colorScheme.surfaceVariant)
-                  else
-                      AssistChipDefaults.assistChipColors(
-                          containerColor = MaterialTheme.colorScheme.secondary),
-              border = null,
-              modifier = Modifier.testTag("FollowButton").padding(5.dp),
-              onClick = {
-                if (associationFollowed.value) {
-                  viewModel.unfollowAssociation(asso.id)
-                } else {
-                  viewModel.followAssociation(asso.id)
-                }
-                profileViewModel.updateNeeded()
-              },
-              label = {
-                if (associationFollowed.value) {
-                  Text(
-                      text = "Following",
-                      color = MaterialTheme.colorScheme.onSurfaceVariant,
-                      fontFamily = FontFamily(Font(R.font.sf_pro_display_regular)),
-                      fontWeight = FontWeight.Medium,
-                  )
-                } else {
-                  Text(
-                      text = "Follow",
-                      color = MaterialTheme.colorScheme.onSecondary,
-                      fontFamily = FontFamily(Font(R.font.sf_pro_display_regular)),
-                      fontWeight = FontWeight.Medium,
-                  )
-                }
-              },
-          )
-        }
-      })
-}
-
-@Composable
-fun JoinUsButton(onClick: () -> Unit, text: String = "Join us") {
-  FloatingActionButton(
-      onClick = onClick,
-      modifier = Modifier.height(42.dp).testTag("JoinButton"),
-      elevation = FloatingActionButtonDefaults.elevation(5.dp),
-      containerColor = MaterialTheme.colorScheme.primary,
-  ) {
-    Text(
-        modifier = Modifier.padding(horizontal = 10.dp),
-        text = text,
-        fontSize = 16.sp,
-        fontFamily = FontFamily(Font(R.font.sf_pro_display_regular)),
-        fontWeight = FontWeight.SemiBold,
-        color = MaterialTheme.colorScheme.onPrimary,
-        style = MaterialTheme.typography.bodyMedium)
   }
 }
